# packages/api-client/src/generated/openapi-types.ts
// Generated by qa/openapi-contract.mjs. Do not edit by hand.

export interface ApiErrorResponse {
  readonly requestId: string
  readonly code: string
  readonly message: string
  readonly status: number
}

export type ApiPath =
  | '/api/admin/global-roles/audit-log'
  | '/api/auth/login'
  | '/api/auth/logout'
  | '/api/channels/{channelId}/messages'
  | '/api/gateway/sessions/{sessionId}/events'
  | '/api/guilds'
  | '/api/guilds/{guildId}/channels'
  | '/api/security/csp-report'
  | '/api/users/@me/guilds'
  | '/api/voice/channels/{channelId}/join'

export type ApiMethod = 'DELETE' | 'GET' | 'PATCH' | 'POST' | 'PUT'

export interface ApiOperation {
  readonly method: ApiMethod
  readonly path: ApiPath
  readonly operationId: string
}

export const apiOperations = [
  {
    "method": "GET",
    "path": "/api/admin/global-roles/audit-log",
    "operationId": "listGlobalRoleAuditLog"
  },
  {
    "method": "POST",
    "path": "/api/auth/login",
    "operationId": "login"
  },
  {
    "method": "POST",
    "path": "/api/auth/logout",
    "operationId": "logout"
  },
  {
    "method": "GET",
    "path": "/api/channels/{channelId}/messages",
    "operationId": "listMessages"
  },
  {
    "method": "POST",
    "path": "/api/channels/{channelId}/messages",
    "operationId": "createMessage"
  },
  {
    "method": "GET",
    "path": "/api/gateway/sessions/{sessionId}/events",
    "operationId": "listGatewayEvents"
  },
  {
    "method": "POST",
    "path": "/api/guilds",
    "operationId": "createGuild"
  },
  {
    "method": "POST",
    "path": "/api/guilds/{guildId}/channels",
    "operationId": "createChannel"
  },
  {
    "method": "POST",
    "path": "/api/security/csp-report",
    "operationId": "submitCspReport"
  },
  {
    "method": "GET",
    "path": "/api/users/@me/guilds",
    "operationId": "listCurrentUserGuilds"
  },
  {
    "method": "POST",
    "path": "/api/voice/channels/{channelId}/join",
    "operationId": "joinVoiceChannel"
  }
] as const satisfies readonly ApiOperation[]